% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BandSelect.R
\name{BandSelect}
\alias{BandSelect}
\title{Bandwidth Selection for the Local Likelihood Estimation of the Dependence Parameter for Right Censored Survival Data}
\usage{
BandSelect(pilot = NULL, nband = 10, u1, u2, status1, status2, X, eta.init,
  family, p = 1, Kern = KernEpa, optim.method = "nlminb",
  control.finite = TRUE, subset.index = 1:n)
}
\arguments{
\item{pilot}{user-specified pilot bandwidth values (optional)}

\item{nband}{desired length of bandwidth values if pilot values are not specified}

\item{u1, u2}{numeric vectors of equal length taking values in [0,1]}

\item{status1, status2}{vectors of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{eta.init}{initial value of calibration parameters for optimization}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{p}{degree of local polynomial in the calibration model: p=0 for local constant, p=1 for local linear}

\item{Kern}{Kernel function}

\item{optim.method}{method used in optimization, current options are "optim" and "nlminb"}

\item{control.finite}{logical indicator to exclude nonfinite loglikelihood contributions (default value = FALSE)}

\item{subset.index}{a vector of indices of data points used in bandwidth selection (optional)}
}
\description{
This function performs bandwidth selection for the local likelihood estimation of the calibration function in conditional copula models for right censored survival data given the conditional margins.
}

