% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beran.R
\name{Beran}
\alias{Beran}
\title{The Beran Estimator}
\usage{
Beran(t, x, Y, status, X, band, Kern = KernEpa)
}
\arguments{
\item{t}{specific time value at which the conditional survival function is calculated}

\item{x}{specific covariate value given which the conditional survival function is calculated}

\item{Y}{vector of event times}

\item{status}{vector of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{band}{bandwidth parameter}

\item{Kern}{Kernel function}
}
\description{
This function defines the Beran estimator to estimate conditional marginal survival functions for right censored event time data given a covariate.
}
\keyword{internal}

