% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeranInv.R
\name{BeranInv}
\alias{BeranInv}
\title{Inverse of the Beran Estimator}
\usage{
BeranInv(u, x, Y, status, X, band, Kern = KernEpa)
}
\arguments{
\item{u}{specific value in [0,1] at which the inverse of the conditional survival function is calculated}

\item{x}{specific covariate value given which the inverse of the conditional survival function is calculated}

\item{Y}{vector of event times}

\item{status}{vector of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{band}{bandwidth parameter}

\item{Kern}{Kernel function}
}
\description{
This function calculates the inverse of the Beran estimator required to generate bootstrap samples for the generalized likelihood ratio test.
}
\keyword{internal}

