% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootSim.Censor.R
\name{BootSim.Censor}
\alias{BootSim.Censor}
\title{Simulating Censoring Times for GLRT Bootstrap Samples}
\usage{
BootSim.Censor(Y1, Y2, status1, status2)
}
\arguments{
\item{Y1, Y2}{vectors of event times}

\item{status1, status2}{vectors of censoring indicators}
}
\description{
This function simulates censoring times in boostrap samples of right censored survival data
}
\keyword{internal}

