% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootSim.NPMar.R
\name{BootSim.NP}
\alias{BootSim.NP}
\title{Simulating Event Time Data for GLRT Bootstrap Samples with Nonparametric Margins}
\usage{
BootSim.NP(Y1, Y2, status1, status2, X, family, cop.par, mar.band,
  Kern = KernEpa)
}
\arguments{
\item{Y1, Y2}{vectors of event times}

\item{status1, status2}{vectors of censoring indicators}

\item{X}{vector of covariate values}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{cop.par}{copula parameter}

\item{mar.band}{vector of bandwidth values for the estimation of conditional margins}

\item{Kern}{Kernel function}
}
\description{
This function simulates right censored event time data for boostrap samples when conditional margins are estimated nonparametrically using the Beran's estimator.
}
\keyword{internal}

