% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootSim.P.R
\name{BootSim.P}
\alias{BootSim.P}
\title{Simulating Event Time Data for GLRT Bootstrap Samples with Parametric Margins}
\usage{
BootSim.P(Y1, Y2, status1, status2, X, family, cop.par, mar.par)
}
\arguments{
\item{Y1, Y2}{vectors of event times}

\item{status1, status2}{vectors of censoring indicators}

\item{X}{vector of covariate values}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{cop.par}{copula parameter}

\item{mar.par}{vector of marginal parameters under the Weibull model. The entries are of the form: (rho1, rho2, lambda1, lambda2, b1, b2)}
}
\description{
This function simulates right censored event time data for boostrap samples when conditional margins are estimated parametrically using a Weibull model.
}
\keyword{internal}

