% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopSurv.Fit.R
\name{CondiCopSurv.Fit}
\alias{CondiCopSurv.Fit}
\title{Fitting a Conditional Copula to Right Censored Survival Data}
\usage{
CondiCopSurv.Fit(u1, u2, status1, status2, X, x, eta.init = NULL, family,
  p = 1, band, Kern = KernEpa, optim.method = "nlminb",
  control.finite = TRUE, update.eta = FALSE, position)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length taking values in [0,1]}

\item{status1, status2}{vectors of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{x}{vector of covariate values at which the estimation is performed}

\item{eta.init}{initial value(s) of calibration parameter for optimization (optional)}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{p}{degree of local polynomial in the calibration model: p=0 for local constant, p=1 for local linear}

\item{band}{bandwidth value used in the conditional copula estimation}

\item{Kern}{Kernel function}

\item{optim.method}{method used in optimization, current options are "optim" and "nlminb"}

\item{control.finite}{logical indicator to exclude nonfinite loglikelihood contributions (default value = FALSE)}

\item{update.eta}{logical indicator to update the initial value of the calibration parameter for optimization}

\item{position}{integer value specifying the starting location for the function estimation}
}
\description{
This function performs the local likelihood estimation of the calibration function in conditional copula models for right censored survival data when the conditional margins are estimated using the Weibull model.
}

