% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopSurv.LocFit.R
\name{CondiCopSurv.LocFit}
\alias{CondiCopSurv.LocFit}
\title{Local Likelihood Estimation of the Dependence Parameter for Right Censored Survival Data}
\usage{
CondiCopSurv.LocFit(u1, u2, status1, status2, X, x, eta.init, family, p = 1,
  band, Kern = KernEpa, optim.method = "nlminb", control.finite = TRUE)
}
\arguments{
\item{u1, u2}{numeric vectors of equal length taking values in [0,1]}

\item{status1, status2}{vectors of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{x}{fixed numeric value in the covariate range}

\item{eta.init}{initial value of calibration parameters for optimization}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{p}{degree of local polynomial in the calibration model: p=0 for local constant, p=1 for local linear}

\item{band}{bandwidth value used in the conditional copula estimation}

\item{Kern}{Kernel function}

\item{optim.method}{method used in optimization; options are "optim" and "nlminb"}

\item{control.finite}{logical indicator to exclude nonfinite loglikelihood contributions (default value = FALSE)}
}
\description{
This function performs the local likelihood estimation of the calibration function in conditional copula models for right censored survival data
}
\keyword{internal}

