% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopSurv.LocLLik.R
\name{CondiCopSurv.LocLLik}
\alias{CondiCopSurv.LocLLik}
\title{Conditional Copula Local Loglikelihood Function}
\usage{
CondiCopSurv.LocLLik(eta, family, p = 1, u1, u2, status1, status2, X, x, band,
  Kern = KernEpa, control.finite = FALSE, negative = TRUE)
}
\arguments{
\item{eta}{calibration parameter}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{p}{degree of local polynomial in the calibration model: p=0 for local constant, p=1 for local linear}

\item{u1, u2}{numeric vectors of equal length taking values in [0,1]}

\item{status1, status2}{vectors of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{x}{fixed numeric value in the covariate range}

\item{band}{bandwidth value used in the conditional copula estimation}

\item{Kern}{Kernel function}

\item{control.finite}{logical indicator to exclude nonfinite loglikelihood contributions (default value = FALSE)}

\item{negative}{logical indicator to return the loglikelihood or negative loglikelihood function (default value = TRUE)}
}
\description{
This function gives the objective function for the local likelihood estimation in conditional copula models for right censored survival data
}
\keyword{internal}

