% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiCopSurv.R
\name{CondiCopSurv}
\alias{CondiCopSurv}
\title{Conditional Copula Model Fitting and Testing for Right Censored Survival Data}
\usage{
CondiCopSurv(Y1, Y2, status1, status2, X, x, family,
  mar.method1 = c("Weibull", "Beran"), mar.method2 = c("Weibull", "Beran"),
  mar.par1, mar.par2, mar.band1, mar.band2, cop.band, mar.pilot = NULL,
  cop.pilot = NULL, nband.mar = 10, nband.cop = 10,
  band.method = c("CV_Y", "CV_T", "CV_use"), eta.init, p = 1,
  Kern = KernEpa, optim.method = "nlminb", control.finite = TRUE,
  update.eta = FALSE, position, return.options = c(par = TRUE, test = TRUE,
  band = TRUE, pval = TRUE), nBoot = 100, boot.options = c(update.marpar =
  TRUE, update.copband = FALSE, update.marband = FALSE), ...)
}
\arguments{
\item{Y1, Y2}{vectors of event times}

\item{status1, status2}{vectors of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{x}{vector of covariate values at which the estimation is performed}

\item{family}{an integer defining the bivariate copula family:

3 = Clayton copula

4 = Gumbel copula

5 = Frank copula}

\item{mar.method1, mar.method2}{method to obtain conditional marginal survival fuctions for each margin: "Weibull" or "Beran"}

\item{mar.par1, mar.par2}{marginal parameters of the Weibull model (optional)}

\item{mar.band1, mar.band2}{bandwidth value used in the Beran estimator (optional)}

\item{cop.band}{bandwidth value used in the conditional copula estimation (optional)}

\item{mar.pilot, cop.pilot}{user-specified pilot bandwidth values (optional)}

\item{nband.mar, nband.cop}{desired length of bandwidth values if pilot values are not specified}

\item{band.method}{method used in bandwidth selection: three cross-validation based selectors}

\item{eta.init}{initial value of calibration parameters for optimization}

\item{p}{degree of local polynomial in the calibration model: p=0 for local constant, p=1 for local linear}

\item{Kern}{Kernel function}

\item{optim.method}{method used in optimization, current options are "optim" and "nlminb"}

\item{control.finite}{logical indicator to exclude nonfinite loglikelihood contributions (default value = TRUE)}

\item{update.eta}{logical indicator to update the initial value of the calibration parameter for optimization}

\item{return.options}{logical vector specifying which results to return}

\item{nBoot}{number of bootstrap samples in the test procedure}

\item{boot.options}{logical vector specifying bootstrap control options}
}
\description{
This function performs the local likelihood estimation of the calibration function in conditional copula models for right censored survival data given a covariate. The function also returns an approximate p-value for the Generalized Likelihood Ratio (GLR) test to assess the significance of the covariate effect.
}

