% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiMarFit.NP.R
\name{CondiMarFit.NP}
\alias{CondiMarFit.NP}
\title{Fitting the Conditional Survival Functions}
\usage{
CondiMarFit.NP(Y, status, X, band, Kern = KernEpa, pilot = NULL,
  nband = 10, band.method = c("CV_Y", "CV_T", "CV_use"), ...)
}
\arguments{
\item{Y}{vector of event times}

\item{status}{vector of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{band}{bandwidth value used in the Beran estimator (optional)}

\item{pilot}{user-specified pilot bandwidth values (optional)}

\item{nband}{desired length of bandwidth values if pilot values are not specified}

\item{band.method}{method used in bandwidth selection: three cross-validation based selectors}
}
\description{
This function fits conditional marginal survival functions for right censored survival data.
}
\keyword{internal}

