% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiMarFit.P.R
\name{CondiMarFit.P}
\alias{CondiMarFit.P}
\title{Fitting the Conditional Survival Functions}
\usage{
CondiMarFit.P(Y, status, X, par)
}
\arguments{
\item{Y}{vector of event times}

\item{status}{vector of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{mar.par}{marginal parameters of the Weibull model: (optional) estimated if missing.}
}
\description{
This function fits conditional marginal survival functions for right censored survival data.
}
\keyword{internal}

