% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondiMarFit.R
\name{CondiMarFit}
\alias{CondiMarFit}
\title{Fitting the Conditional Survival Functions}
\usage{
CondiMarFit(Y, status, X, mar.method = c("Weibull", "Beran"), ...)
}
\arguments{
\item{Y}{vector of event times}

\item{status}{vector of censoring indicators taking values 0 (censored) or 1 (uncensored)}

\item{X}{vector of covariate values}

\item{mar.method}{method to obtain conditional marginal survival fuctions: "Weibull" or "Beran"}
}
\description{
This function fits conditional marginal survival functions for right censored survival data.
}

