% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KernEpa.R
\name{KernEpa}
\alias{KernEpa}
\title{Epanechnikov Kernel}
\usage{
KernEpa(t)
}
\arguments{
\item{t}{numeric vector}
}
\description{
This function defines the Epanechnikov kernel used in the local likelihood estimation.
}
\keyword{internal}

