% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{Null.init}
\alias{Null.init}
\title{Initial Values for Optimization}
\usage{
Null.init(family)
}
\description{
These functions determine initial parameter values for the (local) likelihood estimation.
}
\keyword{internal}

